--- src/main/java/traveler/travel/entity/File.java
@@ -1,18 +1,0 @@
-package traveler.travel.entity;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-import javax.persistence.*;
-
-@Entity
-@Getter
-public class File extends BaseTimeEntity{
-
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "file_id")
-    private Long id;
-    private String originName;
-    private String storedName;
-    private int size;
-}
