+++ src/main/java/traveler/travel/enums/AccountType.java
@@ -1,4 +1,4 @@
-package traveler.travel;
+package traveler.travel.enums;
 
 public enum AccountType {
     EMAIL, KAKAO, GOOGLE, NAVER
