+++ src/main/java/traveler/travel/enums/Authority.java
@@ -1,4 +1,4 @@
-package traveler.travel;
+package traveler.travel.enums;
 
 public enum Authority {
     ROLE_USER, ROLE_ADMIN
