+++ src/main/java/traveler/travel/enums/ChatRoomType.java
@@ -1,4 +1,4 @@
-package traveler.travel;
+package traveler.travel.enums;
 
 public enum ChatRoomType {
     ONE_TO_ONE, GROUP
