+++ src/main/java/traveler/travel/enums/Gender.java
@@ -1,4 +1,4 @@
-package traveler.travel;
+package traveler.travel.enums;
 
 public enum Gender {
     MALE, FEMALE, FREE
