+++ src/main/java/traveler/travel/enums/TravelType.java
@@ -1,4 +1,4 @@
-package traveler.travel;
+package traveler.travel.enums;
 
 public enum TravelType {
     IMMEDIATELY, CONVERSATION
